Relocate Observable method implementations
